+++ cbc/code_test.go
@@ -45,6 +45,10 @@ func TestCode_Validate(t *testing.T) {
 			code: cbc.Code("B3-1-2"),
 		},
 		{
+			name: "valid with slash",
+			code: cbc.Code("B3/12"),
+		},
+		{
 			name: "empty",
 			code: cbc.Code(""),
 		},
@@ -71,6 +75,11 @@ func TestCode_Validate(t *testing.T) {
 		{
 			name:    "lower case",
 			code:    cbc.Code("ab"),
+			wantErr: "valid format",
+		},
+		{
+			name:    "multiple symbols",
+			code:    cbc.Code("AB/.CD"),
 			wantErr: "valid format",
 		},
 		{